datalake: refactor to support custom dlq table schema
But stop right before extending the schema because this type of changes
aren't support by the currently released schema evolution
implementation. After schema merging feature is released we'll be able
to customize the table.

We could add a feature gate but that would be too complex and there is
no urgency in releasing this change so put in on the shelf for now.