ear This is a combination of 2 commits.
cloud_storage: introduce remote_service

We need to create dynamically remote instances to support operations
across buckets that might not share the same underlying client
pool/connections/credentials.

The first step in this direction is to lift out of the remote members
that are expected to be unique, i.e. metrics, materialized resources,
filters. remote_service serves this purpose.